requirements.txt
pytest
requests
python-dotenv
tinytag
fastapi
uvicorn[standard]
streamlit
python-multipart
